refactor: split the stat card widget into its own component
Prepares it for being shared between the new flags in production and the archived to created ration sections.
